Extract more jail dataset utils
